# platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: BrambiLab API
  version: 0.5.0
  description: |
    Base contract for web v1 (docs/architecture/web-v1.md §10).
    WEB-001: operation endpoints and the shared error shape. WEB-002: owner-only
    GitHub OAuth and sessions. WEB-003: private content, translations and
    immutable revisions. WEB-004: media library and authorized delivery under /media.
    WEB-005: publication, durable scheduling and withdrawal per locale; minimal public reader.
    Each later task adds its routes here first.
    Unsafe methods require an Origin (or Referer) equal to PUBLIC_ORIGIN; otherwise
    403 origin_rejected.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: /api/v1
paths:
...
  /media/{assetId}:
    servers: [{ url: / }]
    parameters: [{ $ref: "#/components/parameters/AssetID" }]
    get:
      operationId: getMedia
      summary: Inline image or video. Supports Range (206/416). Resources are only served by /download.
      description: |
        Owner session: any ready asset (private, no-store). Anonymous: only ready + public_enabled
        assets referenced by a published revision of visible content (Cache-Control: no-cache + ETag).
        Anything else is an identical 404 without metadata, also for HEAD and Range.
      tags: [media]
      security: [{}, { session: [] }]
      parameters:
        - { name: Range, in: header, schema: { type: string, example: "bytes=0-1023" } }
      responses:
        "200": { description: "Full content.", content: { "*/*": { schema: { type: string, format: binary } } } }
        "206": { description: "Partial content (Content-Range).", content: { "*/*": { schema: { type: string, format: binary } } } }
        "304": { description: "Not modified (ETag)." }
        "404": { description: "Unknown, private or not inline-servable." }
        "416": { description: "Range not satisfiable or malformed." }
    head:
      operationId: headMedia
      summary: Same authorization and headers as GET, without body.
      tags: [media]
      security: [{}, { session: [] }]
      responses:
        "200": { description: "Headers only." }
        "404": { description: "Unknown or private." }
  /media/{assetId}/download:
    servers: [{ url: / }]
    parameters: [{ $ref: "#/components/parameters/AssetID" }]
    get:
      operationId: downloadMedia
      summary: Attachment download. Anonymous access also requires downloadable.
      tags: [media]
      security: [{}, { session: [] }]
      parameters:
        - { name: Range, in: header, schema: { type: string } }
      responses:
        "200": { description: "Attachment.", content: { "*/*": { schema: { type: string, format: binary } } } }
        "206": { description: "Partial content.", content: { "*/*": { schema: { type: string, format: binary } } } }
        "304": { description: "Not modified." }
        "404": { description: "Unknown, private or not downloadable." }
        "416": { description: "Range not satisfiable." }
    head:
      operationId: headDownload
      summary: Same authorization and headers as GET download, without body.
      tags: [media]
      security: [{}, { session: [] }]
      responses:
        "200": { description: "Headers only." }
        "404": { description: "Unknown, private or not downloadable." }
...
components:
  parameters:
...
    AssetID:
      { name: assetId, in: path, required: true, schema: { type: string, format: uuid } }
...
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: __Host-bl_session
      description: Opaque session cookie (HttpOnly, Secure, SameSite=Lax, Path=/). Named bl_session on http://localhost.
...